+++ storage/src/journal/segmented/fixed.rs
@@ -30,10 +30,10 @@
 //! (and their corresponding blobs) independently.
 
 use super::manager::{AppendFactory, Config as ManagerConfig, Manager};
-use crate::{DestroyPlan, journal::Error};
+use crate::journal::Error;
 use commonware_codec::{CodecFixed, CodecFixedShared, DecodeExt as _, ReadExt as _};
 use commonware_runtime::{
-    Blob, Buf, Handle, Metrics, Storage,
+    Blob, Buf, Handle, Metrics, RemoveTarget, Storage,
     buffer::paged::{CacheRef, Replay as BlobReplay},
 };
 use commonware_utils::NZUsize;
@@ -269,9 +269,14 @@ impl<E: Storage + Metrics, A: CodecFixedShared> Inner<E, A> {
         self.manager.rewind_section(section, size).await
     }
 
-    /// Wait for pending section syncs, then prepare the journal for removal.
-    async fn prepare_destroy(self) -> Result<DestroyPlan<E>, Error> {
-        self.manager.prepare_destroy().await
+    /// Return a context capable of removing this journal's namespace entries.
+    fn destroy_context(&self) -> E {
+        self.manager.destroy_context()
+    }
+
+    /// Wait for pending section syncs, then return the journal's removal targets.
+    async fn into_remove_targets(self) -> Result<Vec<RemoveTarget>, Error> {
+        self.manager.into_remove_targets().await
     }
 
     /// See [Journal::clear].
@@ -576,16 +581,19 @@ impl<E: Storage + Metrics, A: CodecFixedShared> Journal<E, A> {
 
     /// Remove all underlying blobs.
     pub async fn destroy(self) -> Result<(), Error> {
-        self.prepare_destroy()
-            .await?
-            .destroy()
-            .await
-            .map_err(Error::Runtime)
+        let context = self.destroy_context();
+        let targets = self.into_remove_targets().await?;
+        context.remove_batch(targets).await.map_err(Error::Runtime)
     }
 
-    /// Wait for pending section syncs, then prepare the journal for removal.
-    pub(crate) async fn prepare_destroy(self) -> Result<DestroyPlan<E>, Error> {
-        self.0.prepare_destroy().await
+    /// Return a context capable of removing this journal's namespace entries.
+    pub(crate) fn destroy_context(&self) -> E {
+        self.0.destroy_context()
+    }
+
+    /// Wait for pending section syncs, then return the journal's removal targets.
+    pub(crate) async fn into_remove_targets(self) -> Result<Vec<RemoveTarget>, Error> {
+        self.0.into_remove_targets().await
     }
 
     /// Clear all data, resetting the journal to an empty state.
